@@ -1,14 +1,15 @@
 """Functions & classes for compiling derived aspects of the FERC Form 714 data."""
 from functools import cached_property
-from typing import Any, Dict, List
+from typing import Any
 
 import numpy as np
 import pandas as pd
 
 import pudl
 from pudl.metadata.fields import apply_pudl_dtypes
+from pudl.workspace.datastore import Datastore
 
-ASSOCIATIONS: List[Dict[str, Any]] = [
+ASSOCIATIONS: list[dict[str, Any]] = [
     # MISO: Midwest Indep System Operator
     {"id": 56669, "from": 2011, "to": [2009, 2010]},
     # SWPP: Southwest Power Pool
@@ -47,7 +48,7 @@ The changes are applied locally to EIA 861 tables.
   Rows are excluded from `balancing_authority_assn_eia861` with target year and state.
 """
 
-UTILITIES: List[Dict[str, Any]] = [
+UTILITIES: list[dict[str, Any]] = [
     # (no code): Pacific Gas & Electric Co
     {"id": 14328, "reassign": True},
     # (no code): San Diego Gas & Electric Co
@@ -189,7 +190,7 @@ def categorize_eia_code(eia_codes, ba_ids, util_ids, priority="balancing_authori
     return df
 
 
-class Respondents(object):
+class Respondents:
     """A class coordinating compilation of data related to FERC 714 Respondents.
 
     The FERC 714 Respondents themselves are not complex as they are reported, but
@@ -235,14 +236,17 @@ class Respondents(object):
         util_ids=None,
         priority="balancing_authority",
         limit_by_state=True,
+        ds=None,
     ):
         """Set respondent compilation parameters."""
         self.pudl_out = pudl_out
+        self.ds = ds
 
         if pudl_settings is None:
             pudl_settings = pudl.workspace.setup.get_defaults()
         self.pudl_settings = pudl_settings
-
+        if ds is None:
+            ds = Datastore()
         if ba_ids is None:
             ba_ids = (
                 self.balancing_authority_eia861.balancing_authority_id_eia.dropna().unique()
@@ -584,8 +588,7 @@ class Respondents(object):
         """
         if update or self._counties_gdf is None:
             census_counties = pudl.output.censusdp1tract.get_layer(
-                layer="county",
-                pudl_settings=self.pudl_settings,
+                layer="county", pudl_settings=self.pudl_settings, ds=self.ds
             )
             self._counties_gdf = pudl.analysis.service_territory.add_geometries(
                 self.fipsify(update=update), census_gdf=census_counties
