@@ -3,9 +3,10 @@
 import logging
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Any, List, NamedTuple
+from typing import Any, NamedTuple
 from urllib.parse import urlparse
 
+import google.auth
 from google.cloud import storage
 from google.cloud.storage.blob import Blob
 
@@ -112,7 +113,12 @@ class GoogleCloudStorageCache(AbstractCache):
         if parsed_url.scheme != "gs":
             raise ValueError(f"gsc_path should start with gs:// (found: {gcs_path})")
         self._path_prefix = Path(parsed_url.path)
-        self._bucket = storage.Client().bucket(parsed_url.netloc)
+        # Get GCP credentials and billing project id
+        # A billing project is now required because zenodo-cache is requester pays.
+        credentials, project_id = google.auth.default()
+        self._bucket = storage.Client(credentials=credentials).bucket(
+            parsed_url.netloc, user_project=project_id
+        )
 
     def _blob(self, resource: PudlResourceKey) -> Blob:
         """Retrieve Blob object associated with given resource."""
@@ -147,7 +153,7 @@ class LayeredCache(AbstractCache):
     layers are read-only (get).
     """
 
-    def __init__(self, *caches: List[AbstractCache], **kwargs: Any):
+    def __init__(self, *caches: list[AbstractCache], **kwargs: Any):
         """Creates layered cache consisting of given cache layers.
 
         Args:
@@ -155,7 +161,7 @@ class LayeredCache(AbstractCache):
               of decreasing priority.
         """
         super().__init__(**kwargs)
-        self._caches: List[AbstractCache] = list(caches)
+        self._caches: list[AbstractCache] = list(caches)
 
     def add_cache_layer(self, cache: AbstractCache):
         """Adds caching layer. The priority is below all other."""
