@@ -10,13 +10,13 @@ import sqlalchemy as sa
 
 import pudl
 from pudl.convert.censusdp1tract_to_sqlite import censusdp1tract_to_sqlite
+from pudl.workspace.datastore import Datastore
 
 logger = logging.getLogger(__name__)
 
 
 def get_layer(
-    layer: Literal["state", "county", "tract"],
-    pudl_settings=None,
+    layer: Literal["state", "county", "tract"], pudl_settings=None, ds=None
 ) -> gpd.GeoDataFrame:
     """Select one layer from the Census DP1 database.
 
@@ -46,10 +46,12 @@ def get_layer(
         )
     if pudl_settings is None:
         pudl_settings = pudl.workspace.setup.get_defaults()
+    if ds is None:
+        ds = Datastore()
     # Check if we have the Census DP1 database. If not, create it.
     if not Path(pudl_settings["sqlite_dir"], "censusdp1tract.sqlite").exists():
         logger.info("Census DP1 SQLite DB is missing. Creating it.")
-        censusdp1tract_to_sqlite(pudl_settings=pudl_settings)
+        censusdp1tract_to_sqlite(pudl_settings=pudl_settings, ds=ds)
 
     dp1_engine = sa.create_engine(pudl_settings["censusdp1tract_db"])
 
