@@ -38,13 +38,13 @@ def test_no_null_cols_eia(pudl_out_eia, live_dbs, cols, df_name):
 @pytest.mark.parametrize(
     "df_name,raw_rows,monthly_rows,annual_rows",
     [
-        ("bf_eia923", 1_309_942, 1_309_942, 109_807),
+        ("bf_eia923", 1_309_908, 1_309_908, 109_773),
         ("bga_eia860", 117_930, 117_930, 117_930),
-        ("frc_eia923", 560_377, 230_063, 22_686),
-        ("gen_eia923", 559_570, 559_570, 46_718),
+        ("frc_eia923", 560_360, 230_057, 22_679),
+        ("gen_eia923", 559_546, 559_546, 46_694),
         ("gens_eia860", 491_469, 491_469, 491_469),
-        ("gf_eia923", 2_507_870, 2_507_870, 214_709),
-        ("gf_nonuclear_eia923", 2_492_481, 2_492_481, 213_422),
+        ("gf_eia923", 2_507_830, 2_507_830, 214_669),
+        ("gf_nonuclear_eia923", 2_492_441, 2_492_441, 213_382),
         ("gf_nuclear_eia923", 23_498, 23_498, 1_964),
         ("own_eia860", 79_311, 79_311, 79_311),
         ("plants_eia860", 171_570, 171_570, 171_570),
